requirements.txt
# PRRSV病毒衣壳蛋白抑制剂设计项目依赖包

# 核心科学计算包
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 生物信息学和分子处理
biopython>=1.79
rdkit-pypi==2022.9.5
meeko>=0.5.0

# 3D可视化
py3Dmol>=2.0.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
kaleido>=0.2.1

# Web界面
streamlit>=1.20.0

# 机器学习和深度学习
scikit-learn>=1.0.0
torch>=1.10.0

# 数据处理和工具
requests>=2.25.0
tqdm>=4.62.0
pyyaml>=6.0

# 图像处理
pillow>=8.3.0

# 并行计算
joblib>=1.1.0

# 日志和配置
configparser>=5.0.0

# 类型检查
typing-extensions>=4.0.0

# 测试框架
pytest>=6.2.0
pytest-cov>=2.12.0

# 代码质量
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
python-docx>=0.8.11

# 文件处理
openpyxl>=3.0.0
xlrd>=2.0.0
